requirements.txt
torch
numpy
rasterio
matplotlib
tqdm
opencv-python
paddlepaddle
pandas
Pillow
